requirements.txt
fastapi==0.98.0
python-multipart==0.0.6
pydantic==1.10.10
uvicorn==0.22.0
gunicorn==20.1.0
python-dotenv==1.0.0
itsdangerous==2.1.2
setproctitle==1.3.2
jinja2==3.1.2
# indieweb-utils==0.8.1
git+https://github.com/vicchi/indieweb-utils.git@gg-jwt-fix
Pillow==10.0.0
beautifulsoup4==4.10.0
PyJWT==2.4.0
